\name{YPInterimTesting-internal}
\alias{YPInterimTesting-internal}
\alias{fun_estimate}
\alias{fun_interpol}
\alias{awlrstat}
\alias{fun_less}
\title{Internal Functions for Interim Monitoring Using Adaptively Weighted Log-rank Test in Clinical Trials}
\description{
Internal functions for the YPInterimTesting package.
}
\keyword{ypinterim}
